Make algorithm for password_hash configurable.
This allows to use the new Argon2i algorithm introduced in PHP 7.2 (https://wiki.php.net/rfc/argon2_password_hash).

Since this feature only makes sense for PHP 7.2, it's also fine to bump the minimum version requirement to 7.0, remove the not needed compatibility library and upgrade PhpUnit so that the test suite does not throw any warnings or deprecation notices on 7.2.